Hack: If a Field Value is Null instruct Firestore to delete it
This is done by setting the json value of said Field to `FieldValue.delete()`. It instructs Firestore to delete this Field in the collection.

That behaviour is helpful when using `.set(newData, SetOptions(merge: true))`. Which means that the `newData` will be merged into the existing data in the Firestore database.

This way an explicit delete instruction for a specific field is required for it to be deleted.

The model file which wants to use this has to import the related firestore plugin file and has to set `includeIfNull` option to `false`:

        import 'package:cloud_firestore/cloud_firestore.dart';
        ...
        @JsonSerializable(includeIfNull: false)
        class ModelClass { ... }

This implementation of the above functionality is still unfinished. Its still not more than a hack and lacks proper use of an Option of `json_serializable` and so on.